Add digital goods field to dashboard catalogue form
The fact that we are using a NullBooleanField makes this a rather ugly
tri-state drop-down, rather than a simple checkbox. But it's still the
better option, being able to discern which product classes are
actually classified as non-digital goods (those that have the field
set to No), and which haven't been classified at all (those where the
field is None).